feat : 앨범상세화면 댓글 UI 구현.
