+++ src/styles/index.css
@@ -122,8 +122,8 @@ table {
   text-align: right;
   padding: 0 2rem 0.5rem;
 }
-/* 
+
 .layout-wrap .image-gallery-content .image-gallery-slide .image-gallery-image {
   max-width: 100%;
   max-height: none;
-} */
+}
